Cleanup and reorganizing
